cria estrutura de pastas e flag nokill

+++ xcodegen.sh
@@ -1,4 +1,11 @@
-killall Xcode
+#!/bin/bash
+
+if [ "$1" == "nokill" ]; then
+    echo "Skipping killall"
+else
+    killall Xcode
+fi
+
 rm -rf PiraruCooks.xcodeproj 
 xcodegen
 open PiraruCooks.xcodeproj
